Refactor: Split chatbot service into modules
Split the chatbot service into smaller, more focused files to improve maintainability.
